+++ assets/txs.js
@@ -1,14 +1,14 @@
 module.exports = [
   {
     hash: "not a real hash",
-    time: Date.now(), // set by Voyager
+    time: 1042 + 1000, // set by Voyager
     height: "3436",
     tx: {
       type: "8EFE47F0625DE8",
       value: {
         msg: [
           {
-            type: "EAFDE32A2C87F8",
+            type: "cosmos-sdk/Send",
             value: {
               inputs: [
                 {
@@ -79,7 +79,7 @@ module.exports = [
   },
   {
     hash: "not a real hash2",
-    time: Date.now() + 10, // set by Voyager
+    time: 1042 + 1010, // set by Voyager
     height: "3438",
     tx: {
       type: "8EFE47F0625DE8",
@@ -87,7 +87,7 @@ module.exports = [
       value: {
         msg: [
           {
-            type: "EAFDE32A2C87F8",
+            type: "cosmos-sdk/Send",
             value: {
               inputs: [
                 {
@@ -159,14 +159,14 @@ module.exports = [
   {
     hash: "not a real hash3",
 
-    time: Date.now() + 100, // set by Voyager
-    height: "3466",
+    time: 1042 + 100, // set by Voyager
+    height: "466",
     tx: {
       type: "8EFE47F0625DE8",
       value: {
         msg: [
           {
-            type: "EAFDE32A2C87F8",
+            type: "cosmos-sdk/Send",
             value: {
               inputs: [
                 {
@@ -203,9 +203,132 @@ module.exports = [
                       amount: 42
                     }
                   ],
-                  address: "tb1d4u5zerywfjhxuc9nudvw"
+                  address: "tb1da6xsetjg9jxgun9wdesexv05j"
                 }
               ]
+            }
+          }
+        ],
+        fee: {
+          amount: [
+            {
+              denom: "",
+              amount: 0
+            }
+          ],
+          gas: 500000
+        },
+        gas: 500000,
+        signatures: [
+          {
+            pub_key: {
+              type: "AC26791624DE60",
+              value: "70df5mUA3bFTJheAI6umXvABNuUU+q/3nOPS6lmT3tU="
+            },
+            signature: {
+              type: "6BF5903DA1DB28",
+              value:
+                "oODzZz2t1u678Jhy2zFnFbBu7S2IP1zjOIoyVDOvGkQ6WkiiRF3zxEJbNAjHGdDjXf7/NQpMOtpt8FC1x6e6DA=="
+            },
+            account_number: 28,
+            sequence: 0
+          }
+        ]
+      },
+      result: {
+        gasUsed: "3324",
+        tags: [
+          {
+            key: "c2VuZGVy",
+            value: "N0ZFQzVEREE0NDA3RDVBRUQ5OTlFMTM3QjYxQTQ2QTEzQTc2MkJCNg=="
+          },
+          {
+            key: "cmVjaXBpZW50",
+            value: "N0ZFQzVEREE0NDA3RDVBRUQ5OTlFMTM3QjYxQTQ2QTEzQTc2MkJCNg=="
+          }
+        ],
+        fee: {}
+      }
+    }
+  },
+  {
+    hash: "not a real hash4",
+
+    time: 1042 + 150, // set by Voyager
+    height: "566",
+    tx: {
+      type: "8EFE47F0625DE8",
+      value: {
+        msg: [
+          {
+            type: "cosmos-sdk/MsgDelegate",
+            value: {
+              validator_addr:
+                "cosmosvaladdr15ky9du8a2wlstz6fpx3p4mqpjyrm5ctqzh8yqw",
+              delegation: {
+                amount: 4200,
+                denom: "steak"
+              }
+            }
+          }
+        ],
+        fee: {
+          amount: [
+            {
+              denom: "",
+              amount: 0
+            }
+          ],
+          gas: 500000
+        },
+        gas: 500000,
+        signatures: [
+          {
+            pub_key: {
+              type: "AC26791624DE60",
+              value: "70df5mUA3bFTJheAI6umXvABNuUU+q/3nOPS6lmT3tU="
+            },
+            signature: {
+              type: "6BF5903DA1DB28",
+              value:
+                "oODzZz2t1u678Jhy2zFnFbBu7S2IP1zjOIoyVDOvGkQ6WkiiRF3zxEJbNAjHGdDjXf7/NQpMOtpt8FC1x6e6DA=="
+            },
+            account_number: 28,
+            sequence: 0
+          }
+        ]
+      },
+      result: {
+        gasUsed: "3324",
+        tags: [
+          {
+            key: "c2VuZGVy",
+            value: "N0ZFQzVEREE0NDA3RDVBRUQ5OTlFMTM3QjYxQTQ2QTEzQTc2MkJCNg=="
+          },
+          {
+            key: "cmVjaXBpZW50",
+            value: "N0ZFQzVEREE0NDA3RDVBRUQ5OTlFMTM3QjYxQTQ2QTEzQTc2MkJCNg=="
+          }
+        ],
+        fee: {}
+      }
+    }
+  },
+  {
+    hash: "not a real hash5",
+
+    time: 1042 + 150, // set by Voyager
+    height: "566",
+    tx: {
+      type: "8EFE47F0625DE8",
+      value: {
+        msg: [
+          {
+            type: "cosmos-sdk/BeginUnbonding",
+            value: {
+              validator_addr:
+                "cosmosvaladdr15ky9du8a2wlstz6fpx3p4mqpjyrm5ctqzh8yqw",
+              shares: 100
             }
           }
         ],