+++ src/components/TmLiAnyTransaction/index.js
@@ -1,0 +1,5 @@
+import TmLiAnyTransaction from "./TmLiAnyTransaction.vue"
+
+export default function install(Vue) {
+  Vue.component("tm-li-any-transaction", TmLiAnyTransaction)
+}