+++ src/components/TmLiTransaction/tranaction-colors.js
@@ -1,0 +1,11 @@
+export default {
+  bank: {
+    sent: "#25293F",
+    received: "#215C2F"
+  },
+  stake: {
+    bonded: "#512584", // and redelegation
+    unbonded: "#9D5C19",
+    redelegate: "#7227B1"
+  }
+}