+++ src/index.js
@@ -104,6 +104,15 @@ export {
   default as TmLiTransaction
 } from "./components/TmLiTransaction/TmLiTransaction.vue"
 export {
+  default as TmLiBankTransaction
+} from "./components/TmLiBankTransaction/TmLiBankTransaction.vue"
+export {
+  default as TmLiStakeTransaction
+} from "./components/TmLiStakeTransaction/TmLiStakeTransaction.vue"
+export {
+  default as TmLiAnyTransaction
+} from "./components/TmLiAnyTransaction/TmLiAnyTransaction.vue"
+export {
   default as TmModalError
 } from "./components/TmModalError/TmModalError.vue"
 export { default as TmModule } from "./components/TmModules/TmModule.vue"